% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles-package.r, R/bubbles.R
\docType{package}
\name{bubbles}
\alias{bubbles}
\alias{bubbles-package}
\title{bubbles.}
\usage{
bubbles(value, label, key = NULL, tooltip = "", color = "#EEEEEE",
  textColor = "#333333", textSize = 8, width = NULL, height = NULL)
}
\arguments{
\item{value}{Numeric vector of values, used for sizing bubbles. The value 
will be proportional to the radius, not the area; for area, call 
\code{\link{sqrt}} on the value.}

\item{label}{Character vector of textual labels to use on bubbles}

\item{key}{Values that will uniquely identify a bubble across runs. This
doesn't matter for static bubble charts, but if a bubble chart receives
updates (i.e. in a Shiny app) then d3.js will use the key to know which
bubbles in the "before" state correspond with which bubbles in the
"after" state, and perform smooth transitions. If \code{NULL}, then the
key will effectively be the row number.}

\item{tooltip}{Character vector of tooltip values, to be shown on hover}

\item{color}{Character vector of bubble color values, in \code{"#RRGGBB"}
format; can be length 1 or length of \code{value}}

\item{textColor}{Character vector of text color values, in \code{"#RRGGBB"}
format; can be length 1 or length of \code{value}}

\item{textSize}{Character vector of text size values; can be length 1 or 
length of \code{value}}

\item{width}{The width of the widget, or \code{NULL} for default.}

\item{height}{The height of the widget, or \code{NULL} for default.}
}
\value{
A widget object that can be printed at the console, outputted as part
  of an R Markdown document, or used in a Shiny app with 
  \code{\link{renderBubbles}}.
}
\description{
Creates a bubble chart.
}
\examples{
bubbles(runif(10), LETTERS[1:10], color = rainbow(10, alpha = NULL))
  
}
