% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bubbles.R
\name{bubblesOutput}
\alias{bubblesOutput}
\alias{renderBubbles}
\title{Bubble chart functions for Shiny}
\usage{
bubblesOutput(outputId, width = "600px", height = "600px")

renderBubbles(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Shiny output ID for this bubble chart}

\item{width,height}{Plot width/height. Must be a valid CSS unit (e.g. "100%",
"400px") or a number, which will be interpreted as pixels ("px").}

\item{expr}{An expression that generates a \code{\link{bubbles}} widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})?}
}
\description{
Bubble chart functions for Shiny
}

