% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/bubbles-package.r, R/bubbles.R
\docType{package}
\name{bubbles}
\alias{bubbles}
\alias{bubbles-package}
\title{bubbles.}
\usage{
bubbles(value, label, tooltip = NULL, color = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{value}{Numeric vector of values, used for sizing bubbles}

\item{label}{Character vector of textual labels to use on bubbles}

\item{tooltip}{Character vector of tooltip values, to be shown on hover}

\item{color}{Character vector of color values, in \code{"#RRGGBB"} format}

\item{width}{The width of the widget, or \code{NULL} for default.}

\item{height}{The height of the widget, or \code{NULL} for default.}
}
\value{
A widget object that can be printed at the console, outputted as part
  of an R Markdown document, or used in a Shiny app with
  \code{\link{renderBubbles}}.
}
\description{
bubbles.

Creates a bubble chart.
}

